Rename region to non-interactable-surface
Based on feedback, renamed the `region` instances to
non-interactable-surface to be more clear.
