+++ packages/hyperion-react-testapp/src/App.tsx
@@ -3,7 +3,6 @@
  */
 
 import React, { ChangeEventHandler, useCallback, useState } from 'react';
-import logo from './logo.svg';
 import './App.css';
 import LargeComp from './component/LargeComponent';
 import Counter from "./component/Counter";
@@ -16,7 +15,7 @@ import RecursiveRuncComponent from "./component/RecursiveFuncComponent";
 import { ElementTextTooltip } from "@hyperion/hyperion-autologging-visualizer/src/component/ElementTextTooltip.react";
 import { SyncChannel } from './Channel';
 import { ALSessionGraph } from "@hyperion/hyperion-autologging-visualizer/src/component/ALSessionGraph.react";
-import RegionComponent from './component/RegionComponent';
+import NonInteractiveSurfaceComponent from './component/NonInteractiveSurfaceComponent';
 
 function InitComp() {
   const [count, setCount] = React.useState(0);
@@ -40,7 +39,7 @@ function App() {
   const maxDepth = 1000;
 
   const Modes = {
-    'region': () => <RegionComponent></RegionComponent>,
+    'mutationOnlySurface': () => <NonInteractiveSurfaceComponent></NonInteractiveSurfaceComponent>,
     'network': () => <DynamicSvgComponent></DynamicSvgComponent>,
     'nested': () => <ElementTextTooltip channel={SyncChannel}>
       <div>
@@ -56,17 +55,15 @@ function App() {
       <div>
         <ElementNameComponent />
       </div>
-      <DynamicSvgComponent></DynamicSvgComponent>
       <TextComponent />
       <RecursiveRuncComponent i={3}></RecursiveRuncComponent>
-      <RegionComponent></RegionComponent>
     </ElementTextTooltip>,
   };
-  const [mode, setMode] = useState<keyof typeof Modes>('region');
+  const [mode, setMode] = useState<keyof typeof Modes>('mutationOnlySurface');
 
   const onChange = useCallback<ChangeEventHandler<HTMLSelectElement>>((event) => {
     const value = event.target.value;
-    if (value === 'region' || value === 'network' || value === 'nested') {
+    if (value === 'mutationOnlySurface' || value === 'network' || value === 'nested') {
       setMode(value);
     }
   }, []);
